Re-organize e2e tests
