--- src/cnv_upgrade_utilities/utils.py
@@ -1,5 +1,0 @@
-"""Backward-compatible re-exports. Import from specific modules for new code."""
-
-from cnv_upgrade_utilities.post_upgrade_suites import *  # noqa: F401,F403
-from cnv_upgrade_utilities.upgrade_types import *  # noqa: F401,F403
-from cnv_upgrade_utilities.version_types import *  # noqa: F401,F403
