+++ tox.ini
@@ -39,10 +39,10 @@ commands =
     pytest tests/e2e/ -m e2e --log-cli-level=INFO {posargs}
 
 [testenv:generate]
-description = Generate upgrade path and release checklist snapshots (requires RH network)
+description = Generate current testing paths (requires RH network)
 passenv = VERSION_EXPLORER_URL
 commands =
-    python scripts/generate_upgrade_snapshots.py {posargs}
+    python scripts/generate_current_testing_paths.py {posargs}
 
 [testenv:coverage]
 description = Report coverage and enforce threshold
